lunchvote/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.0,<6.0
django-cors-headers>=3.0,<4.0  # If you're handling cross-origin requests

